review sivun koodit refaktoroitu showreviews ja writereview komponentteihin. pitää testata vielä lisää
